remove().UUID

+++ database/migrations/2020_01_01_000006_create_contacts_table.php
@@ -15,8 +15,8 @@ class CreateContactsTable extends Migration
     {
         Schema::create('contacts', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('account_id')->index();
-            $table->integer('organization_id')->nullable()->index();
+            $table->bigInteger('account_id')->index();
+            $table->bigInteger('organization_id')->nullable()->index();
             $table->string('first_name', 25);
             $table->string('last_name', 25);
             $table->string('email', 50)->nullable();
